# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sound2Soul API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Creator profiles
...
  /tracks:
    get:
      operationId: getMyTracks
      tags: [tracks]
      summary: Get creator's own tracks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
    post:
      operationId: createTrack
      tags: [tracks]
      summary: Upload a new track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tracks/{slug}:
    get:
      operationId: getTrackBySlug
      tags: [tracks]
      summary: Get a public track by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrack
      tags: [tracks]
      summary: Update own track
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
    delete:
      operationId: deleteTrack
      tags: [tracks]
      summary: Delete own track
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reports:
    post:
      operationId: reportTrack
      tags: [reports]
      summary: Report a track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Waitlist
...
  /admin/tracks:
    get:
      operationId: getAdminTracks
      tags: [admin]
      summary: Get all tracks for admin
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTrack"
  /admin/tracks/{trackId}/approve:
    post:
      operationId: approveTrack
      tags: [admin]
      summary: Approve a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/tracks/{trackId}/reject:
    post:
      operationId: rejectTrack
      tags: [admin]
      summary: Reject a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectTrackBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/tracks/{trackId}/feature:
    post:
      operationId: featureTrack
      tags: [admin]
      summary: Feature or unfeature a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureTrackBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/tracks/{trackId}:
    delete:
      operationId: adminDeleteTrack
      tags: [admin]
      summary: Admin delete a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [admin]
      summary: Get all reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, reviewed, resolved]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /admin/reports/{reportId}/resolve:
    post:
      operationId: resolveReport
      tags: [admin]
      summary: Resolve a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveReportBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SignupBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [creator, listener]
      required: [email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [creator, listener, admin]
        createdAt:
          type: string
      required: [id, email, role, createdAt]
    UserWithProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [creator, listener, admin]
        createdAt:
          type: string
        creatorProfile:
          $ref: "#/components/schemas/CreatorProfile"
      required: [id, email, role, createdAt]
    CreatorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        artistName:
          type: string
        slug:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        genres:
          type: array
          items:
            type: string
        moodIdentityTags:
          type: array
          items:
            type: string
        aiToolsUsed:
          type: array
          items:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        creatorStatement:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, artistName, slug]
    CreatorPublicProfile:
      type: object
      properties:
        id:
          type: string
        artistName:
          type: string
        slug:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        genres:
          type: array
          items:
            type: string
        moodIdentityTags:
          type: array
          items:
            type: string
        aiToolsUsed:
          type: array
          items:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        creatorStatement:
          type: string
        followerCount:
          type: integer
        trackCount:
          type: integer
        createdAt:
          type: string
      required: [id, artistName, slug]
...
    Track:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
        isFeatured:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        saveCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorPublicProfile"
      required: [id, creatorId, title, slug, visibility, moderationStatus]
    TrackDetail:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
        isFeatured:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        saveCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorPublicProfile"
        isLiked:
          type: boolean
        isSaved:
          type: boolean
      required: [id, creatorId, title, slug, visibility, moderationStatus]
    CreateTrackBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
      required: [title, audioUrl, visibility]
    UpdateTrackBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
...
    ReportBody:
      type: object
      properties:
        trackId:
          type: string
        reason:
          type: string
          enum: [copyright_concern, unauthorized_voice, offensive_content, misleading_disclosure, other]
        details:
          type: string
        reporterEmail:
          type: string
      required: [trackId, reason]
...
    AdminTrack:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        aiInvolvementType:
          type: string
        visibility:
          type: string
        moderationStatus:
          type: string
        isFeatured:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        reportCount:
          type: integer
        createdAt:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorPublicProfile"
      required: [id, title, slug, visibility, moderationStatus]
...
    Report:
      type: object
      properties:
        id:
          type: string
        trackId:
          type: string
        trackTitle:
          type: string
        trackSlug:
          type: string
        reporterEmail:
          type: string
        reason:
          type: string
        details:
          type: string
        status:
          type: string
          enum: [open, reviewed, resolved]
        createdAt:
          type: string
      required: [id, trackId, reason, status, createdAt]
    RejectTrackBody:
      type: object
      properties:
        reason:
          type: string
    FeatureTrackBody:
      type: object
      properties:
        featured:
          type: boolean
      required: [featured]
    ResolveReportBody:
      type: object
      properties:
        status:
          type: string
          enum: [reviewed, resolved]
      required: [status]